Implement FrozenDict ourselves
