+++ python/pyiceberg/typedef.py
@@ -15,15 +15,18 @@
 # specific language governing permissions and limitations
 # under the License.
 
-from typing import Dict, Tuple
+from typing import Any, Dict, Tuple
 
 
 class FrozenDict(Dict):
-    def __set__(self, instance, value):
-        raise AttributeError("FrozenSet does not support set")
+    def __setitem__(self, instance, value):
+        raise AttributeError("FrozenDict does not support assignment")
+
+    def update(self, *args: Any, **kwargs: Any) -> None:  # type: ignore
+        raise AttributeError("FrozenDict does not support .update()")
 
 
-EMPTY_DICT = frozendict()
+EMPTY_DICT = FrozenDict()
 
 Identifier = Tuple[str, ...]
 Properties = Dict[str, str]
